requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
python-multipart==0.0.20
opencv-python==4.11.0.86
numpy==2.2.2
pdf2image==1.17.0
Pillow==11.1.0
pillow-heif==0.21.0
imutils==0.5.4
rapidocr-onnxruntime==1.4.4
